requirements.txt
anyio==3.6.2
asgiref==3.4.1
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
fastapi==0.83.0
h11==0.13.0
idna==3.10
immutables==0.19
importlib-metadata==4.8.3
pydantic==1.9.2
python-dotenv==0.20.0
requests==2.27.1
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.1.1
urllib3==1.26.20
uvicorn==0.16.0
zipp==3.6.0
